requirements.txt
litellm>=1.40.0
groq>=0.9.0
arxiv>=2.1.0
pypdf>=4.2.0
python-dotenv>=1.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.6.0
numpy>=1.26.0
streamlit>=1.32.0
rank-bm25>=0.2.2
ragas>=0.1.5
cohere>=5.0.0
plotly>=5.18.0
pytest>=8.0.0
datasets>=2.17.0
langchain-openai>=0.0.8
